The T-pipe operator is similar to the pipe operator, but it returns the value of the left-hand side expression instead of the result of the right-hand side function. It is useful for side effects, such as printing or plotting. See %T>% for details.

The T-pipe operator is similar to the pipe operator, but it returns the value of the left-hand side expression instead of the result of the right-hand side function. It is useful for side effects, such as printing or plotting. See %T>% for details.

Usage

lhs %T>% rhs

Arguments

lhs

A value or the magrittr placeholder.

rhs

A function call using the magrittr semantics.

Value

The value of lhs.